Modify scripts to build relative to scritps
Previously, the scripts were built relative to the working
directory. The problem with that is that I was creating
symbolic links and then having to change directories when
re-running scripts. This change always builds the files
relative to the script directory. The use will be to
clone this repo into the directory where you want to build
things.